requirements/dev.txt
################################################
## DO NOT EDIT - AUTOMATICALLY GENERATED FILE ##
################################################
astroid==1.4.5
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
click==3.3
cloverpy==0.1.1
colorama==0.3.7
cov-core==1.15.0
coverage==4.0.3
docopt==0.6.2
docutils==0.12
git+https://github.com/catapultbamboo/file_translate.git@0.0.3
Flask==0.10.1
funcsigs==1.0.2
grip==3.2.0
html5lib==0.9999999
isodate==0.5.4
itsdangerous==0.24
Jinja2==2.8
keepalive==0.5
lazy-object-proxy==1.2.2
Markdown==2.6.6
MarkupSafe==0.23
mock==1.3.0
opp-bass==1.0.9
opp-chub==1.0.5
opp-koi==1.0.8
path-and-address==1.1.0
pbr==1.9.1
py==1.4.31
Pygments==2.1.3
PyJWT==1.4.0
pylint==1.5.5
pyparsing==2.1.1
pytest==2.8.7
pytest-cov==1.8.0
python-dateutil==2.4.2
rdflib==4.2.1
rdflib-jsonld==0.3
rdflib-sparql==0.2
requests==2.4.3
singledispatch==3.4.0.3
six==1.10.0
SPARQLWrapper==1.7.6
Sphinx==1.2.3
tornado==4.3
Werkzeug==0.11.9
wheel==0.24.0
wrapt==1.10.8
